flask_app/requirements.txt
attrs==23.1.0
blinker==1.6.2
click==8.1.7
exceptiongroup==1.1.3
flasgger==0.9.7.1
Flask==2.3.3
Flask-Testing==0.8.1
gunicorn==21.2.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
mistune==3.0.1
packaging==23.1
pluggy==1.3.0
psycopg2-binary==2.9.7
pytest==7.4.2
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.2
six==1.16.0
tomli==2.0.1
Werkzeug==2.3.7
